+++ backend/app/src/main/java/com/ugent/pidgeon/postgre/repository/GroupClusterRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface GroupClusterRepository extends JpaRepository<GroupClusterEntity, Long> {
-    List<GroupClusterEntity> findByCourseId(int courseId);
+    List<GroupClusterEntity> findByCourseId(long courseId);
 }
